Create separate prop-types.js file
